working_vendora/requirements_cloud.txt
# VENDORA Cloud Platform Requirements
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google Cloud Platform
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.16.4
google-cloud-firestore==2.13.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Firebase (optional for enhanced auth)
firebase-admin==6.2.0

# AI and ML
google-generativeai==0.3.2

# Data processing
pandas==2.1.4
numpy==1.24.4

# Async and utilities
asyncio
aiofiles==23.2.1
httpx==0.25.2
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0

# Security
python-jose[cryptography]==3.3.0
cryptography==41.0.8

# Environment management
python-dotenv==1.0.0

# Optional: Enhanced monitoring
prometheus-client==0.19.0